+++ .vscode/launch.json
@@ -5,7 +5,7 @@
 	"version": "0.2.0",
 	"configurations": [
 		{
-			"name": "gcc - Build and debug active file",
+			"name": "clang - Build and debug active file",
 			"type": "cppdbg",
 			"request": "launch",
 			"program": "${workspaceRoot}/minishell",
@@ -14,21 +14,8 @@
 			"cwd": "${fileDirname}",
 			"environment": [],
 			"externalConsole": false,
-			"MIMode": "gdb",
-			"setupCommands": [
-				{
-					"description": "Enable pretty-printing for gdb",
-					"text": "-enable-pretty-printing",
-					"ignoreFailures": true
-				},
-				{
-					"description": "Set Disassembly Flavor to Intel",
-					"text": "-gdb-set disassembly-flavor intel",
-					"ignoreFailures": true
-				}
-			],
-			"preLaunchTask": "C/C++: gcc build active file",
-			"miDebuggerPath": "/usr/bin/gdb"
+			"MIMode": "lldb",
+			"preLaunchTask": "C/C++: clang build active file"
 		}
 	]
 }
